pyproject.toml
[tool.poetry]
name = "systemlink-store-archiver"
version = "1.0"
description = "A salt execution module for archiving the contents of the SystemLink Store and Forward store directory"
authors = []

[tool.poetry.dependencies]
python = "^3.6.2"
salt = "^3004"
python-dateutil = "^2.8.2"
systemlink-sdk = "^21.1.0"
boto3 = "~1.15.18"

[tool.poetry.dev-dependencies]
# Lint
flake8 = ">=3.8.4"
flake8-docstrings = ">=1.5.0"
flake8-import-order = ">=0.18.1"
pep8-naming = ">=0.11.1"
black = ">=20.8b1"

# Test
pytest = ">=6.0"

[[tool.poetry.source]]
name = "NI PyPI"
url = "https://pypi.ni.systems"
default = true

[tool.black]
line-length = 100

[build-system]
requires = [ "poetry-core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"
